macros: Split up the crate into modules
