+++ charon/macros/Cargo.toml
@@ -8,7 +8,5 @@ edition = "2021"
 proc-macro = true
 
 [dependencies]
-syn = "1.0.102"
 quote = "1.0.21"
-serde = { version = "1.0.145", features = ["derive"] }
-toml = "0.5.9"
+syn = "1.0.102"
