+++ charon/tests/ui/issue-4-slice-try-into-array.rs
@@ -1,0 +1,3 @@
+pub fn trait_error(s: &[u8]) {
+    let _array: [u8; 4] = s.try_into().unwrap();
+}
